refactor(core): source: rename screen source to media projection source
